api/requirements.txt
ruff==0.14.14

fastapi>=0.128.0

uvicorn[standard]==0.40.0
aiosqlite
pydantic
starlette
httpx
pytest
python-dateutil